✏️ Change to use the new design abstraction of AI model provider.
